hermes-api/requirements.txt
numpy
requests
python-dotenv
